Move `ForCachingExtendedStatisticsAccess` to out package
